# apps/web/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateDecisionRequest": {
        "properties": {
          "content": {
            "additionalProperties": true,
            "description": "A JDM decision graph.",
            "type": "object"
          },
          "key": {
            "description": "Lowercase words with single hyphens; folders separated by `/`.",
            "maxLength": 120,
            "minLength": 1,
            "pattern": "^[a-z][a-z0-9]*(-[a-z0-9]+)*(/[a-z][a-z0-9]*(-[a-z0-9]+)*)*$",
            "type": "string"
          }
        },
        "required": [
          "key"
        ],
        "type": "object"
      },
...
      "DecisionListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DecisionSummaryResponse"
            },
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "type": "object"
      },
...
      "DecisionResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/DecisionSummaryResponse"
          },
          {
            "properties": {
              "content": {
                "additionalProperties": true,
                "description": "A JDM decision graph.",
                "type": "object"
              }
            },
            "required": [
              "content"
            ],
            "type": "object"
          }
        ]
      },
      "DecisionSummaryResponse": {
        "properties": {
          "changedSinceVersion": {
            "description": "Whether the draft has changed since that version (true before the first).",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "key": {
            "description": "How graphs call this decision, e.g. `bureau/normalize`.",
            "type": "string"
          },
          "latestVersion": {
            "description": "The number of its latest version; absent before the first \"Save version\".",
            "format": "int32",
            "type": [
              "integer",
              "null"
            ]
          },
          "revision": {
            "description": "Grows by one on every save; send it back when saving.",
            "format": "int32",
            "type": "integer"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedBy": {
            "$ref": "#/components/schemas/PersonRef"
          }
        },
        "required": [
          "id",
          "key",
          "revision",
          "updatedAt",
          "updatedBy",
          "changedSinceVersion"
        ],
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "description": "Stable, machine-readable code, e.g. `DECISION_NOT_FOUND`.",
            "type": "string"
          },
          "details": {
            "description": "Engine error document (failing node and reason) for evaluation failures."
          },
          "fields": {
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            },
            "type": [
              "object",
              "null"
            ]
          },
          "message": {
            "description": "Safe to show to a person.",
            "type": "string"
          },
          "requestId": {
            "description": "Same value as the `x-request-id` response header; quote it when reporting a problem.",
            "type": "string"
          }
        },
        "required": [
          "message",
          "code",
          "requestId"
        ],
        "type": "object"
      },
...
      "PersonRef": {
        "properties": {
          "email": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email"
        ],
        "type": "object"
      },
...
      "SaveDecisionRequest": {
        "properties": {
          "content": {
            "additionalProperties": true,
            "description": "A JDM decision graph.",
            "type": "object"
          },
          "revision": {
            "description": "The revision this draft was loaded or last saved at.",
            "format": "int32",
            "type": "integer"
          }
        },
        "required": [
          "content",
          "revision"
        ],
        "type": "object"
      },
...
      "SimulateDecisionRequest": {
        "properties": {
          "content": {
            "additionalProperties": true,
            "description": "A JDM decision graph.",
            "type": "object"
          },
          "context": {
            "description": "Input passed to the decision.",
            "type": "object"
          }
        },
        "type": "object"
      },
      "SimulateRequest": {
        "description": "Design-time evaluation from the editor. The editor sends the whole draft\nproject so sub-decision references resolve exactly as they will in a release.",
        "properties": {
          "context": {
            "description": "Input passed to the decision.",
            "type": "object"
          },
          "decisions": {
            "description": "Every decision of the project (JDM documents), keyed as graphs reference them.",
            "type": "object"
          },
          "key": {
            "description": "The decision to evaluate.",
            "type": "string"
          }
        },
        "required": [
          "decisions",
          "key"
        ],
        "type": "object"
      },
      "SimulateResponse": {
        "properties": {
          "performance": {
            "description": "Engine evaluation time, e.g. `412.5µs`.",
            "type": "string"
          },
          "result": {
            "type": "object"
          },
          "trace": {
            "description": "Per-node trace: inputs, outputs and timing of each node.",
            "type": "object"
          }
        },
        "required": [
          "result",
          "trace",
          "performance"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Control plane of Donka: authoring, simulation, releases.",
    "license": {
      "name": ""
    },
    "title": "Donka Studio API",
    "version": "unknown"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/projects/{project_id}/decisions": {
      "get": {
        "operationId": "list",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionListResponse"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "PROJECT_NOT_FOUND"
          }
        },
        "summary": "The project's decisions, by key (any member).",
        "tags": [
          "decisions"
        ]
      },
      "post": {
        "operationId": "create",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResponse"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid key (INVALID_REQUEST)"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Viewers cannot create decisions (FORBIDDEN)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "PROJECT_NOT_FOUND"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "DECISION_KEY_TAKEN or PROJECT_ARCHIVED"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The content is not a decision model (INVALID_DECISION)"
          }
        },
        "summary": "A new decision, empty unless content is given (editors and owners).",
        "tags": [
          "decisions"
        ]
      }
    },
    "/projects/{project_id}/decisions/{decision_id}": {
      "delete": {
        "operationId": "delete",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "decision_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Viewers cannot delete (FORBIDDEN)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "PROJECT_NOT_FOUND or DECISION_NOT_FOUND"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "PROJECT_ARCHIVED"
          }
        },
        "summary": "Deletes a decision (editors and owners): it leaves the project and its key\nis free again; its versions stay in the history.",
        "tags": [
          "decisions"
        ]
      },
      "get": {
        "operationId": "get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "decision_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResponse"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "PROJECT_NOT_FOUND or DECISION_NOT_FOUND"
          }
        },
        "summary": "One decision with its draft (any member).",
        "tags": [
          "decisions"
        ]
      },
      "put": {
        "operationId": "save",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "decision_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionSummaryResponse"
                }
              }
            },
            "description": "Saved; the new revision"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Viewers cannot save (FORBIDDEN)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "PROJECT_NOT_FOUND or DECISION_NOT_FOUND"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "DECISION_CONFLICT (details: revision, updatedAt, updatedBy) or PROJECT_ARCHIVED"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The content is not a decision model (INVALID_DECISION)"
          }
        },
        "summary": "Saves the draft (editors and owners). Send the revision the draft was loaded\nat: when someone saved since, the answer is `409 DECISION_CONFLICT` with who\nand when in `details`, and nothing is overwritten.",
        "tags": [
          "decisions"
        ]
      }
    },
    "/projects/{project_id}/decisions/{decision_id}/simulate": {
      "post": {
        "operationId": "simulate",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "decision_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulateDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulateResponse"
                }
              }
            },
            "description": "Result and per-node trace"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "PROJECT_NOT_FOUND or DECISION_NOT_FOUND"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A decision of the project is not a valid model (INVALID_DECISION), or the evaluation failed (EVALUATION_FAILED, trace in details)"
          }
        },
        "summary": "Evaluates this decision with every other decision of the project, as a\nrelease would, and returns the result with a per-node trace (any member:\nviewers simulate too). Nothing is saved.",
        "tags": [
          "decisions"
        ]
      }
    },
...
    "/simulate": {
      "post": {
        "operationId": "simulate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulateResponse"
                }
              }
            },
            "description": "Result and per-node trace"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Malformed request body (INVALID_REQUEST)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No decision with that key (DECISION_NOT_FOUND)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid decision model (INVALID_DECISION) or evaluation error (EVALUATION_FAILED)"
          }
        },
        "tags": [
          "simulator"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ]
}